requirements.txt
stravalib==1.4.1
pandas==2.1.4
python-dotenv==1.0.0
polyline==2.0.2
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
